fix(continuity 1-forms): fix 1-form cont
There was an issue where if neighbouring elements were rotated continuity was not correctly handled. This arose from the fact that 1-forms have "direction".
An example is that if the bottom side of element 1 ended up agains the bottom or left side of element 2. Both of these have the field coming "into" the element, meaning the coefficients in continuity relations for one of the elements must have an opposite sign to those if it were adjacent to the right or top side, where the field is comming "out" of the element.