fixed local refinement
There was an issue caused by strong boundary conditions being too high
order. The reason it can not work is that the child elements can have
higher combined order than their parent. In that case, the number of
equations/Lagrange multipliers can not exceed the order of the parent
element, as otherwise the constraints can not be satisfied and the
problem has no solutions.