final solution to the element question
This better be the last and final time I have to fucking bother with this crap. I am so sick of changing this back and forth. I still need to thoroughly test this, as I am more than sure there will be PLEANTY of errors, but screw that, it's 23.50 and I need sleep!