more refinement control
Added refinement limit for the refinement order, at which point, h-refinement must happen. The error and h-ref cost estimates are now returned with the solution when refinement is done.

Projector calculation code also had to be updated. I still need to check if that's all working correctly now, but I think it's good.