updated forms
The k-forms underwent some changes. This should make them a bit easier to use and more consistent.

First, I got rid of the Hodge. It was a tricky operation, which is not really needed in the FEM operation, since weight functions are used to form inner products.

In line with this, I also changed the interior products. Since there's no more Hodge and I actually derived it in the thesis, the interior products of dual forms should now be instead written as applied to the weight function. There were some sign changes, so that should be double checked. This also allowed to remove the weird behaviour where the interior product operators would either be primal -> dual or dual -> primal operators. Now with removal of the KHodge, everything is just primal and simple. Some C-side code still should be removed, but after that is done, it is finished.

Interior products had the option to compute what I called the "adjoint" operators, which would give the partial with respect to the vector field. This could be used to make a Newton-Raphson solver, but in practice, it is inferior to Picart iterations, so no more
of that.

As a consequence of that, I rewrote the parsing of the k-form expressions into bytecode. It is much nicer. I also followed that up with a new system to string translation function. It no longer does it on the non-translated system, so what it prints is exactly what is getting evaluated.

Since MatMul operation was no longer used, since the Transpose operation was removed quite some time ago, due to not being needed anymore. It is still in the C-code, but Python no longer has it.

Lastly, I changed operators that are used by k-forms. Before I was adding them as I went on, but the result was more and more confusing syntax as time went on. Now `@` is the inner product, `*` the interior product, and `^` the boundary term. No more exceptions or double meanings to these.